+++ store/iavl/tree_test.go
@@ -10,19 +10,12 @@ import (
 
 func TestImmutableTreePanics(t *testing.T) {
 	t.Parallel()
-	db := dbm.NewMemDB()
-	immTree := iavl.NewImmutableTree(db, 100, false)
+	immTree := iavl.NewImmutableTree(dbm.NewMemDB(), 100, false)
 	it := &immutableTree{immTree}
 	require.Panics(t, func() { it.Set([]byte{}, []byte{}) })
 	require.Panics(t, func() { it.Remove([]byte{}) })
 	require.Panics(t, func() { it.SaveVersion() })           // nolint:errcheck
 	require.Panics(t, func() { it.DeleteVersion(int64(1)) }) // nolint:errcheck
-	require.Panics(t, func() { it.LoadVersionForOverwriting(int64(1)) })
-	require.Equal(t, it.AvailableVersions(), []int{})
-
-	val, err := it.GetVersioned(nil, 1)
-	require.Error(t, err)
-	require.Nil(t, val)
 
 	val, proof, err := it.GetVersionedWithProof(nil, 1)
 	require.Error(t, err)