+++ types/store_test.go
@@ -4,8 +4,9 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/suite"
-
 	dbm "github.com/tendermint/tm-db"
+
+	"github.com/line/ostracon/libs/log"
 
 	"github.com/line/lbm-sdk/store/rootmulti"
 	"github.com/line/lbm-sdk/store/types"
@@ -25,7 +26,7 @@ func (s *storeTestSuite) SetupSuite() {
 }
 
 func (s *storeTestSuite) TestPrefixEndBytes() {
-	var testCases = []struct {
+	testCases := []struct {
 		prefix   []byte
 		expected []byte
 	}{
@@ -48,7 +49,7 @@ func (s *storeTestSuite) TestCommitID() {
 	var empty sdk.CommitID
 	s.Require().True(empty.IsZero())
 
-	var nonempty = sdk.CommitID{
+	nonempty := sdk.CommitID{
 		Version: 1,
 		Hash:    []byte("testhash"),
 	}
@@ -109,7 +110,7 @@ func (s *storeTestSuite) TestDiffKVStores() {
 
 func (s *storeTestSuite) initTestStores() (types.KVStore, types.KVStore) {
 	db := dbm.NewMemDB()
-	ms := rootmulti.NewStore(db)
+	ms := rootmulti.NewStore(db, log.NewNopLogger())
 
 	key1 := types.NewKVStoreKey("store1")
 	key2 := types.NewKVStoreKey("store2")