+++ x/authz/generic_authorization.go
@@ -4,9 +4,7 @@ import (
 	sdk "github.com/line/lbm-sdk/types"
 )
 
-var (
-	_ Authorization = &GenericAuthorization{}
-)
+var _ Authorization = &GenericAuthorization{}
 
 // NewGenericAuthorization creates a new GenericAuthorization object.
 func NewGenericAuthorization(msgTypeURL string) *GenericAuthorization {