+++ x/gov/client/rest/rest_test.go
@@ -54,7 +54,7 @@ func (s *IntegrationTestSuite) TestLegacyGetAllProposals() {
 				s.Require().NoError(val.ClientCtx.LegacyAmino.UnmarshalJSON(respJSON, &errResp))
 				s.Require().Equal(errResp.Error, tc.expErrMsg)
 			} else {
-				var resp = rest.ResponseWithHeight{}
+				resp := rest.ResponseWithHeight{}
 				err = val.ClientCtx.LegacyAmino.UnmarshalJSON(respJSON, &resp)
 				s.Require().NoError(err)
 
@@ -105,7 +105,7 @@ func (s *IntegrationTestSuite) TestLegacyGetVote() {
 				s.Require().NoError(val.ClientCtx.LegacyAmino.UnmarshalJSON(respJSON, &errResp))
 				s.Require().Equal(errResp.Error, tc.expErrMsg)
 			} else {
-				var resp = rest.ResponseWithHeight{}
+				resp := rest.ResponseWithHeight{}
 				err = val.ClientCtx.LegacyAmino.UnmarshalJSON(respJSON, &resp)
 				s.Require().NoError(err)
 
@@ -153,7 +153,7 @@ func (s *IntegrationTestSuite) TestLegacyGetVotes() {
 
 				s.Require().Equal(errResp.Error, tc.expErrMsg)
 			} else {
-				var resp = rest.ResponseWithHeight{}
+				resp := rest.ResponseWithHeight{}
 				err = val.ClientCtx.LegacyAmino.UnmarshalJSON(respJSON, &resp)
 				s.Require().NoError(err)