+++ x/ibc/core/23-commitment/types/commitment_test.go
@@ -3,11 +3,14 @@ package types_test
 import (
 	"testing"
 
+	"github.com/stretchr/testify/suite"
+	dbm "github.com/tendermint/tm-db"
+
+	"github.com/line/ostracon/libs/log"
+
 	"github.com/line/lbm-sdk/store/iavl"
 	"github.com/line/lbm-sdk/store/rootmulti"
 	storetypes "github.com/line/lbm-sdk/store/types"
-	"github.com/stretchr/testify/suite"
-	dbm "github.com/tendermint/tm-db"
 )
 
 type MerkleTestSuite struct {
@@ -20,7 +23,7 @@ type MerkleTestSuite struct {
 
 func (suite *MerkleTestSuite) SetupTest() {
 	db := dbm.NewMemDB()
-	suite.store = rootmulti.NewStore(db)
+	suite.store = rootmulti.NewStore(db, log.NewNopLogger())
 
 	suite.storeKey = storetypes.NewKVStoreKey("iavlStoreKey")